Change control flow to use an init function
To make the etch-a-sketch refresh when the user
inputs a new grid dimension, an init function
is now responsible for adding event listeners to
each pixel.
This will make it possible to reinitialize the
drawing grid once the user-defined resolution
feature is working.

makePixelGrid(n) now accepts a resolution argument
corresponding to resolution.