Add grid resolution
User can input a resolution, as the number of
pixels along an edge (i.e. '5' = 5x5 resolution)
and grid will reset accordingly.

To keep the overall grid dimension the same,
the styling has changed. flex-grow is set to 1 for
the columns and the individual pixels to make them
fill the given space.